The size property of MuCircularProgress should be Double

+++ core/src/main/scala/chandu0101/scalajs/react/components/materialui/MuiCircularProgress.scala
@@ -25,7 +25,7 @@ case class MuiCircularProgress(
 for when there is no value for progress.*/
   mode:       js.UndefOr[DeterminateIndeterminate] = js.undefined,
   /* The size of the progress.*/
-  size:       js.UndefOr[Int]                      = js.undefined,
+  size:       js.UndefOr[Double]                      = js.undefined,
   /* Override the inline-styles of the root element.*/
   style:      js.UndefOr[CssProperties]            = js.undefined,
   /* The value of progress, only works in determinate mode.*/
